Moved PemLabel enum inside types.
