+++ src/types.ts
@@ -1,5 +1,13 @@
 export type SignatureAlgorithm = "sha1" | "sha256" | "sha512";
 
+export const PemLabel = {
+  CERTIFICATE: "CERTIFICATE" as const,
+  PUBLIC_KEY: "PUBLIC KEY" as const,
+  PRIVATE_KEY: "PRIVATE KEY" as const,
+};
+
+export type PemLabelId = typeof PemLabel[keyof typeof PemLabel];
+
 export interface SamlSigningOptions {
   privateKey: string | Buffer;
   signatureAlgorithm?: SignatureAlgorithm;
